requirements.txt
fastapi
uvicorn
jinja2
pydantic
qrcode
reportlab
python-barcode
sqlalchemy
psycopg2